--- fullver/backend/src/main/java/com/github/chore/repository/entity/order/OrderRepository.java
@@ -1,9 +1,0 @@
-//package com.github.chore.repository.entity.order;
-//
-//
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface OrderRepository extends JpaRepository<Order,Integer> {
-//}
